LPS-66933 Move opensocial data handlers to the proper packages
