+++ modules/apps/opensocial/opensocial-portlet/docroot/WEB-INF/liferay-portlet.xml
@@ -5,8 +5,8 @@
 	<portlet>
 		<portlet-name>1</portlet-name>
 		<icon>/icon.png</icon>
-		<portlet-data-handler-class>com.liferay.opensocial.admin.lar.AdminPortletDataHandler</portlet-data-handler-class>
-		<staged-model-data-handler-class>com.liferay.opensocial.admin.lar.GadgetStagedModelDataHandler</staged-model-data-handler-class>
+		<portlet-data-handler-class>com.liferay.opensocial.admin.exportimport.data.handler.AdminPortletDataHandler</portlet-data-handler-class>
+		<staged-model-data-handler-class>com.liferay.opensocial.admin.exportimport.data.handler.GadgetStagedModelDataHandler</staged-model-data-handler-class>
 		<control-panel-entry-category>apps</control-panel-entry-category>
 		<header-portlet-css>/admin/css/main.css</header-portlet-css>
 		<css-class-wrapper>opensocial-portlet-admin</css-class-wrapper>
