+++ modules/apps/opensocial/opensocial-portlet/docroot/WEB-INF/src/com/liferay/opensocial/admin/exportimport/data/handler/AdminPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.opensocial.admin.lar;
+package com.liferay.opensocial.admin.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.BasePortletDataHandler;
 import com.liferay.exportimport.kernel.lar.DataLevel;
